tweak clocking parameters -- maybe marginally better?
1) forgot to reset feedback phase to 0 from a previous experiment

2) set the frequency range to the 1080p for filter/lock computations

actually, probably in the long term it would be good to write
down the filter/lock info for the two modes, so we can jam them
into the registers when also setting the PLL. Basically, when
changing the frequency range, the filter/lock parameters also
change automatically when compiling from Vivado, but they don't
share the formula to do it yourself. So you have to set the VCO,
Vivado derives the frequency range, and then derives filter/lock
bits -- then you read those back, create a lookup table of
filter/lock and then use those to optimize the PLL performance.