change the genlock method from pulse to wholesale signal change
Also add a shit ton of comments so I can follow the code.

Instead of taking a start of frame pulse, genlock is now the
full set of hsync, vsync, and de signals extracted from the
incoming frame. This allows correct location of the pixels
even with non-compliant transmitters that are using weird
video modes or erronously implemented video modes.