output/hdmi: fix incorrect clock domain in S7HDMIOutEncoderSerializer
"pix_o" and "pix5x_o" clock domains being used here are incorrect,
since these domains are declared in the HDMI input section of the
core and hence are related to the HDMI input signal. If there is no
input connected then obviously these clocks will not work, and the PLL
generating these will also not lock. So, this is not a correct behaviour
Ideally, these clock domains should be separately declared in the HDMI
output domain and derived from the main system clock.